+++ src/components/layout/SideBar.vue
@@ -8,7 +8,7 @@ import { cmpWorkByTitle } from 'src/lib/work.ts';
 const workStore = useWorkStore();
 
 import { useGoalStore } from 'src/stores/goal.ts';
-import { cmpGoalByProgress } from 'src/lib/goal.ts';
+import { cmpGoalByCompletion } from 'src/lib/goal.ts';
 const goalStore = useGoalStore();
 
 import { useBoardStore } from 'src/stores/board.ts';
@@ -51,7 +51,7 @@ const items = computed(() => {
       to: { name: 'goals' },
       header: true,
     },
-    ...(goalStore.starredGoals ?? []).toSorted(cmpGoalByProgress).map(goal => ({
+    ...(goalStore.starredGoals ?? []).toSorted(cmpGoalByCompletion).map(goal => ({
       key: `goal-${goal.id}`,
       label: goal.title,
       to: { name: 'goal', params: { goalId: goal.id } },