+++ src/components/pages/goals/GoalsListPage.vue
@@ -6,7 +6,7 @@ import { useGoalStore } from 'src/stores/goal.ts';
 const goalStore = useGoalStore();
 
 import { getGoals, GoalWithAchievement } from 'src/lib/api/goal.ts';
-import { cmpGoalByProgress } from 'src/lib/goal.ts';
+import { cmpGoalByCompletion } from 'src/lib/goal.ts';
 
 import ApplicationLayout from 'src/layouts/ApplicationLayout.vue';
 import type { MenuItem } from 'primevue/menuitem';
@@ -47,7 +47,7 @@ const reloadGoals = async function() {
 
 const goalsFilter = ref<string>('');
 const filteredGoals = computed(() => {
-  const sortedGoals = goals.value.toSorted(cmpGoalByProgress);
+  const sortedGoals = goals.value.toSorted(cmpGoalByCompletion);
   const searchTerm = goalsFilter.value.toLowerCase();
   return sortedGoals.filter(goal => goal.title.toLowerCase().includes(searchTerm) || goal.description.toLowerCase().includes(searchTerm));
 });